refactor: move BlockType::read_and_validate to validation
