+++ src/core/structure/types.rs
@@ -3,11 +3,7 @@ use core::fmt;
 use alloc::vec::Vec;
 
 use crate::{
-    assert_validated::UnwrapValidatedExt,
-    core::{
-        decoding::reader::WasmReader,
-        structure::modules::indices::{IdxVec, TypeIdx},
-    },
+    core::structure::modules::indices::{IdxVec, TypeIdx},
     ValidationError,
 };
 
@@ -72,24 +68,6 @@ pub enum BlockType {
 }
 
 impl BlockType {
-    pub fn read_and_validate(
-        wasm: &mut WasmReader,
-        c_types: &IdxVec<TypeIdx, FuncType>,
-    ) -> Result<Self, ValidationError> {
-        if wasm.peek_u8()? == 0x40 {
-            // Empty block type
-            let _ = wasm.read_u8().unwrap_validated();
-            Ok(BlockType::Empty)
-        } else if let Ok(val_ty) = wasm.handle_transaction(|wasm| ValType::read(wasm)) {
-            // No parameters and given valtype as the result
-            Ok(BlockType::Returns(val_ty))
-        } else {
-            // An index to a function type
-            let index = wasm.read_var_i33_as_u32()?;
-            TypeIdx::validate(index, c_types).map(BlockType::Type)
-        }
-    }
-
     /// Converts this block type to a specific [`FuncType`].
     ///
     /// A vector of function types is required, in case the current block type
