+++ src/validation/types.rs
@@ -1,0 +1,31 @@
+use crate::{
+    assert_validated::UnwrapValidatedExt,
+    core::{
+        decoding::reader::WasmReader,
+        structure::{
+            modules::indices::{IdxVec, TypeIdx},
+            types::BlockType,
+        },
+    },
+    FuncType, ValType, ValidationError,
+};
+
+impl BlockType {
+    pub fn read_and_validate(
+        wasm: &mut WasmReader,
+        c_types: &IdxVec<TypeIdx, FuncType>,
+    ) -> Result<Self, ValidationError> {
+        if wasm.peek_u8()? == 0x40 {
+            // Empty block type
+            let _ = wasm.read_u8().unwrap_validated();
+            Ok(BlockType::Empty)
+        } else if let Ok(val_ty) = wasm.handle_transaction(|wasm| ValType::read(wasm)) {
+            // No parameters and given valtype as the result
+            Ok(BlockType::Returns(val_ty))
+        } else {
+            // An index to a function type
+            let index = wasm.read_var_i33_as_u32()?;
+            TypeIdx::validate(index, c_types).map(BlockType::Type)
+        }
+    }
+}
